backend/addToFirestoreRecipeAutoGeneratedByAI/requirements.txt
beautifulsoup4==4.12.3
firebase_admin==6.5.0
functions_framework==3.8.1
protobuf==5.28.2
python-dotenv==1.0.1
Requests==2.32.3
typing_extensions==4.12.2
google-generativeai==0.8.1
